Move client side tests so they're co-located
Moved all client side tests so they're closer to the relevant pieces.
- didn't do it for the backend since most of those tests are straight-forwards
