+++ client/src/components/specs/Body.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import Body from '../../components/Body';
+import Body from '../Body';
 
 describe('A Body component', () => {
   afterEach(() => cleanup())
